Javadoc
* Documented AudioController using javadoc.
* Restructured the directory.
