+++ core/src/com/dylan773/finalyearproject/screens/windows/OptionWindow.java
@@ -1,0 +1,4 @@
+package com.dylan773.finalyearproject.screens.windows;
+
+public class OptionWindow {
+}
