+++ desktop/src/com/dylan773/finalyearproject/desktop/DesktopLauncher.java
@@ -2,7 +2,7 @@ package com.dylan773.finalyearproject.desktop;
 
 import com.badlogic.gdx.backends.lwjgl.LwjglApplication;
 import com.badlogic.gdx.backends.lwjgl.LwjglApplicationConfiguration;
-import com.dylan773.finalyearproject.EducationGame;
+import com.dylan773.finalyearproject.utilities.EducationGame;
 
 public class DesktopLauncher {
 	public static void main (String[] arg) {
@@ -10,6 +10,9 @@ public class DesktopLauncher {
 		config.title = "Edu Quiz";
 		config.height = 720;
 		config.width = 1280;
+		// TODO - not sure if this actually needed idfk
+		//config.useGL30 = true;
+
 		new LwjglApplication(new EducationGame(), config);
 	}
 }
